Plot Rain Data as Bars

Plot Rain Data as Bars

Usage

plotRain(
  timestamps,
  values,
  gaugeName = "",
  xlim = NULL,
  ylim = NULL,
  signal.width = NA,
  shift.to.begin = 0,
  col = "blue",
  pch = 16,
  reverse.ylim = TRUE,
  las = 1,
  innerMargins = c(0, 0, 0, 0),
  add.time.axis = TRUE,
  ...
)

Arguments

timestamps

timestamps

values

precipitation values in mm

gaugeName

Name of rain gauge to appear in the y label

xlim

x limits

ylim

y limits

signal.width

width of time intervals represented by timestamps, in seconds. If NA (default) this value is deduced from the time differences in timestamps

shift.to.begin

value in seconds by which the timstamps need to be shifted in order to get the begin of the time intervals that the timestamps represent. Default: 0 (meaning that the timestamps represent the beginnings of the time intervals). A value of signal.width would mean that the timestamps represent the ends of the time intervals and a value of signal.width/2 would mean that the timestamps represent the midpoints of the time intervals that they represent.

col

fill colour of bars. Default: "blue"

pch

plot character to be used for plotting data points (at the given timestamps), additionally to plotting bars. Set to NA in order to suppress plotting these points

reverse.ylim

if TRUE (default), the y axis is reversed, i.e. the bars are heading from top to bottom

las

numeric in 0,1,2,3; the style of axis labels. See help for par.

innerMargins

passed to plot_variable

add.time.axis

passed to plot_variable

...

further arguments passed to points