Plot a Variable of a Data Frame

Plot a Variable of a Data Frame

Usage

plot_variable(
  hydraulicData,
  variableName = names(hydraulicData)[[2]],
  type = .defaultPlotParameter("type", variableName),
  col = .defaultPlotParameter("col", variableName),
  ylab = .defaultPlotParameter("ylab", variableName),
  pch = .defaultPlotParameter("pch", variableName),
  xlim = NULL,
  ylim = NULL,
  add = FALSE,
  plot.grid = TRUE,
  innerMargins = c(0, 0, 0, 0),
  absolute = FALSE,
  reverse.ylim = FALSE,
  add.time.axis = TRUE,
  time.axis.in.margins = FALSE,
  ...
)

Arguments

hydraulicData

data frame with at least two columns. First column is expected to contain the timestamps

variableName

default: name of second column

type

plot type, passed to points

col

colour, passed to points

ylab

y label, passed to plot

pch

plot character, passed to points

xlim

x limits, passed to plot

ylim

y limits, passed to plot

add

if TRUE points are added to an existing plot

plot.grid

if TRUE a grid is drawn

innerMargins

margins (bottom, left, top, right) within the plot area

absolute

if TRUE, innerMargins are to be interpreted as absolute values instead of fractions of the plot region. Default: FALSE

reverse.ylim

if TRUE (default), the y axis is reversed, i.e. the bars are heading from top to bottom

add.time.axis

if TRUE a time axis is drawn

time.axis.in.margins

if TRUE the time axis is drawn only within the plot area excluding the margins

...

additional arguments passed to plot