Create argument list for mybarplot

Create argument list for mybarplot

args_mybarplot(cex.axis = graphics::par("cex.axis"),
  cex.names = graphics::par("cex.axis"), ..., args.text = NULL,
  args.yaxis = NULL)

Arguments

cex.axis
character expansion factor of axis labels

cex.names
character expansion factor of bar group labels

…
additional arguments to barplot

args.text
list of arguments that are given to text being called to plot the bar group labels. Use args_text to create such a list.

args.yaxis
list of arguments that are given to axis being called to plot the y axis. Use args_yaxis to create such a list.